refactor: split & rename munsellConversions.ts
just moving code & updating imports, no actual changes.
now named colorConversions.ts, with the detection algorithm moved to
colorDetection.ts, and type defs moved to types.ts
