+++ dev-client/src/screens/SoilScreen/ColorScreen/components/ColorDisplay.tsx
@@ -26,10 +26,10 @@ import {
   View,
 } from 'terraso-mobile-client/components/NativeBaseAdapters';
 import {
-  MunsellColor,
   munsellToRGB,
   munsellToString,
-} from 'terraso-mobile-client/model/color/munsellConversions';
+} from 'terraso-mobile-client/model/color/colorConversions';
+import {MunsellColor} from 'terraso-mobile-client/model/color/types';
 
 type Props = {
   color: MunsellColor;
